requirements.txt
python-binance==1.0.18
python-telegram-bot==20.7
python-dotenv==1.0.1
requests==2.31.0
